Rename class to more accurate name
